Moved tests into specific classes.
Cleaned up code
